Remove refactor
