+++ docs/source/internal.mdx
@@ -48,10 +48,6 @@ The main work on your PyTorch `DataLoader` is done by the following function:
 
 [[autodoc]] state.DistributedType
 
-## Multi-Processing Operators
-
-[[autodoc]] operations.gather
-
 ## Tracking
 
 [[autodoc]] tracking.GeneralTracker
@@ -63,6 +59,8 @@ The main work on your PyTorch `DataLoader` is done by the following function:
 [[autodoc]] utils.send_to_device
 
 [[autodoc]] utils.set_seed
+
+[[autodoc]] utils.gather
 
 [[autodoc]] utils.synchronize_rng_state
